moved notes and categories to separated files
