+++ task1/src/data/categories.js
@@ -1,0 +1,10 @@
+import Category from "../models/Category.js";
+
+//categories data
+const categories = [
+    new Category("Task", "cart-fill"),
+    new Category("Random Thought", "lightbulb-fill"),
+    new Category("Idea", "lightning")
+];
+
+export default categories
